WIP - generic useragent refactoring.
