+++ codegen/aws/core/src/main/java/software/amazon/smithy/python/aws/codegen/AwsUserAgentIntegration.java
@@ -8,6 +8,7 @@ import java.util.List;
 import software.amazon.smithy.codegen.core.Symbol;
 import software.amazon.smithy.codegen.core.SymbolReference;
 import software.amazon.smithy.python.codegen.ConfigProperty;
+import software.amazon.smithy.python.codegen.GenerationContext;
 import software.amazon.smithy.python.codegen.integrations.PythonIntegration;
 import software.amazon.smithy.python.codegen.integrations.RuntimeClientPlugin;
 import software.amazon.smithy.utils.SmithyInternalApi;
@@ -18,35 +19,40 @@ import software.amazon.smithy.utils.SmithyInternalApi;
 @SmithyInternalApi
 public class AwsUserAgentIntegration implements PythonIntegration {
     @Override
-    public List<RuntimeClientPlugin> getClientPlugins() {
-        return List.of(
-                RuntimeClientPlugin.builder()
-                        .addConfigProperty(
-                                ConfigProperty.builder()
-                                        .name("user_agent_extra")
-                                        .documentation("Additional suffix to be added to the User-Agent header.")
-                                        .type(Symbol.builder().name("str").build())
-                                        .nullable(true)
-                                        .build())
-                        .addConfigProperty(
-                                ConfigProperty.builder()
-                                        .name("sdk_ua_app_id")
-                                        .documentation(
-                                                "A unique and opaque application ID that is appended to the User-Agent header.")
-                                        .type(Symbol.builder().name("str").build())
-                                        .nullable(true)
-                                        .build())
-                        .pythonPlugin(
-                                SymbolReference.builder()
-                                        .symbol(Symbol.builder()
-                                                .namespace(
-                                                        AwsPythonDependency.SMITHY_AWS_CORE.packageName() + ".plugins",
-                                                        ".")
-                                                .name("user_agent_plugin")
-                                                .addDependency(AwsPythonDependency.SMITHY_AWS_CORE)
-                                                .build())
-                                        .build())
-                        .build());
+    public List<RuntimeClientPlugin> getClientPlugins(GenerationContext context) {
+        if (context.applicationProtocol().isHttpProtocol()) {
+            return List.of(
+                    RuntimeClientPlugin.builder()
+                            .addConfigProperty(
+                                    ConfigProperty.builder()
+                                            .name("user_agent_extra")
+                                            .documentation("Additional suffix to be added to the User-Agent header.")
+                                            .type(Symbol.builder().name("str").build())
+                                            .nullable(true)
+                                            .build())
+                            .addConfigProperty(
+                                    ConfigProperty.builder()
+                                            .name("sdk_ua_app_id")
+                                            .documentation(
+                                                    "A unique and opaque application ID that is appended to the User-Agent header.")
+                                            .type(Symbol.builder().name("str").build())
+                                            .nullable(true)
+                                            .build())
+                            .pythonPlugin(
+                                    SymbolReference.builder()
+                                            .symbol(Symbol.builder()
+                                                    .namespace(
+                                                            AwsPythonDependency.SMITHY_AWS_CORE.packageName()
+                                                                    + ".plugins",
+                                                            ".")
+                                                    .name("aws_user_agent_plugin")
+                                                    .addDependency(AwsPythonDependency.SMITHY_AWS_CORE)
+                                                    .build())
+                                            .build())
+                            .build());
+        } else {
+            return List.of();
+        }
     }
 
 }
