+++ codegen/core/src/main/java/software/amazon/smithy/python/codegen/ClientGenerator.java
@@ -76,7 +76,7 @@ final class ClientGenerator implements Runnable {
             var defaultPlugins = new LinkedHashSet<SymbolReference>();
 
             for (PythonIntegration integration : context.integrations()) {
-                for (RuntimeClientPlugin runtimeClientPlugin : integration.getClientPlugins()) {
+                for (RuntimeClientPlugin runtimeClientPlugin : integration.getClientPlugins(context)) {
                     if (runtimeClientPlugin.matchesService(context.model(), service)) {
                         runtimeClientPlugin.getPythonPlugin().ifPresent(defaultPlugins::add);
                     }
@@ -657,7 +657,7 @@ final class ClientGenerator implements Runnable {
     private void initializeHttpAuthParameters(PythonWriter writer) {
         var derived = new LinkedHashSet<DerivedProperty>();
         for (PythonIntegration integration : context.integrations()) {
-            for (RuntimeClientPlugin plugin : integration.getClientPlugins()) {
+            for (RuntimeClientPlugin plugin : integration.getClientPlugins(context)) {
                 if (plugin.matchesService(context.model(), service)
                         && plugin.getAuthScheme().isPresent()
                         && plugin.getAuthScheme().get().getApplicationProtocol().isHttpProtocol()) {
@@ -746,7 +746,7 @@ final class ClientGenerator implements Runnable {
 
         var defaultPlugins = new LinkedHashSet<SymbolReference>();
         for (PythonIntegration integration : context.integrations()) {
-            for (RuntimeClientPlugin runtimeClientPlugin : integration.getClientPlugins()) {
+            for (RuntimeClientPlugin runtimeClientPlugin : integration.getClientPlugins(context)) {
                 if (runtimeClientPlugin.matchesOperation(context.model(), service, operation)) {
                     runtimeClientPlugin.getPythonPlugin().ifPresent(defaultPlugins::add);
                 }
