+++ packages/smithy-http/src/smithy_http/plugins/__init__.py
@@ -1,9 +1,8 @@
 #  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #  SPDX-License-Identifier: Apache-2.0
 
-from typing import Any
+from smithy_core.interfaces.config import Config
+from smithy_http.interceptors.user_agent import UserAgentInterceptor
 
-from smithy_aws_core.interceptors.user_agent import UserAgentInterceptor
-
-def aws_user_agent_plugin(config: Any) -> None:
+def user_agent_plugin(config: Config) -> None:
     config.interceptors.append(UserAgentInterceptor())
